add additional tests and merge test classes
